Bug 1756800 - [sort frameworks in alert view]
