+++ .gitignore
@@ -40,3 +40,5 @@ coverage.xml
 
 # Docs generation
 docs.egg-info
+
+.python-version
